+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 
 * 修复 **user.args** 通用注入
+* 增加 默认别名 **@tpl/helper**
+* 修改 **workflow.limitImgSize** 为 **workflow.limitResourcesSize**
 
 ## 0.0.19 (May 4, 2018)